% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_daily.R
\name{cbr_currency_ts}
\alias{cbr_currency_ts}
\title{Dynamics of currency over the period with 1-to-1 rate to RUB}
\source{
\url{http://www.cbr.ru/scripts/XML_dynamic.asp?date_req1=02/03/2001&date_req2=14/03/2001&VAL_NM_RQ=R01235; http://www.cbr.ru/scripts/xml_metall.asp?date_req1=01/07/2001&date_req2=13/07/2001}
}
\usage{
cbr_currency_ts(
  currency = "USD",
  from = as.character(Sys.Date() - 30),
  to = as.character(Sys.Date())
)
}
\arguments{
\item{currency}{Currency in ISO 3-char code ('USD', 'GBP', 'EUR', etc.)}

\item{from}{Date in format "yyyy-mm-dd". You also can get tomorrow exchange rates.}

\item{to}{Date in format "yyyy-mm-dd". You also can get tomorrow exchange rates.}
}
\value{
numeric vector of exchange rates with length 1 or more
}
\description{
Dynamics of currency over the period with 1-to-1 rate to RUB
}
