% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_daily.R
\name{cbr_daily}
\alias{cbr_daily}
\title{Daily/monthly Central Bank of Russia exchange currency rates or
precious metal discount prices to russian ruble with 1-to-1 rate}
\source{
\url{http://www.cbr.ru/scripts/XML_daily.asp?date_req=02/03/2002;     http://www.cbr.ru/scripts/xml_metall.asp?date_req1=01/07/2001}
}
\usage{
cbr_daily(currency = "USD", date = TRUE)
}
\arguments{
\item{currency}{Currency in ISO 3-char code ('USD', 'GBP', 'EUR', etc.), or precious metal chemical abbreviation}

\item{date}{Date in format "yyyy-mm-dd". You also can get tomorrow exchange rates.}
}
\value{
numeric 1 length vector with currency of precious metal rates
}
\description{
Daily/monthly Central Bank of Russia exchange currency rates or
precious metal discount prices to russian ruble with 1-to-1 rate
}
