% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_daily.R
\name{cbr_inflation}
\alias{cbr_inflation}
\title{NOTE: function under construction}
\usage{
cbr_inflation(date = Sys.Date(), check_actual = FALSE)
}
\arguments{
\item{date}{date in 'yyyy-mm-dd' format}
}
\value{

}
\description{
NOTE: function under construction
}
