% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_daily.R
\name{cbr_metal_ts}
\alias{cbr_metal_ts}
\title{Dynamics of currency over the period with 1-to-1 rate to RUB}
\source{
\url{https://www.cbr.ru/scripts/xml_metall.asp?date_req1=01/07/2001&date_req2=13/07/2001}
}
\usage{
cbr_metal_ts(
  currency = "ALL",
  from = as.character(Sys.Date() - 30),
  to = as.character(Sys.Date())
)
}
\arguments{
\item{currency}{Precious metals (case insensitive) chemical code c('Au', 'Ag', 'Pt', 'Pd', 'all')}

\item{from}{Date in format "yyyy-mm-dd". You also can get tomorrow exchange rates.}

\item{to}{Date in format "yyyy-mm-dd". You also can get tomorrow exchange rates.}
}
\value{
data frame of date, buy and sell rates of precious metals in long format
}
\description{
Dynamics of currency over the period with 1-to-1 rate to RUB
}
