% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrCurrency-package.R
\docType{package}
\name{cbrCurrency}
\alias{cbrCurrency}
\title{What your package does}
\description{
Longer description of the package goes here.
Probably with links to other resources about it, citations, etc.
}
