% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbr_daily.R
\name{key_rate_ts}
\alias{key_rate_ts}
\title{Return CBR key rate value on date of time series for the period}
\source{
\url{https://www.cbr.ru/hd_base/KeyRate/}
}
\usage{
key_rate_ts(date = Sys.Date(), from = FALSE)
}
\arguments{
\item{date}{Date in format "yyyy-mm-dd". You also can get tomorrow exchange rates.}

\item{from}{Optional date argument for time series return}
}
\value{
numeric vector of length 1 or more
}
\description{
Return CBR key rate value on date of time series for the period
}
